Added check for presence of "getResponse" method
The reason for a rejected Guzzle call was expected to be of a certain
type that had a defined getResponse method. This was incorrect in some
instances and a fatal exception was being thrown. I added a check for
the method and am setting response to null if it is not defined.